employeeUpdate

+++ hrapp/src/main/java/hrlovecraft/Contact.java
@@ -51,7 +51,7 @@ public class Contact {
         return phoneNumber;
     }
 
-    public void setPhoneNumber(int phoneNumber) {
+    public void setPhoneNumber(long phoneNumber) {
         this.phoneNumber = phoneNumber;
     }
